Started work on line camera
A virtual camera that follows the line. A key press could switch between planning robot movement, and player/robot following camera.

Implemented Cinemachine to make the camera feel more fluid. I chose to allow for close zoom in when the player is close to the robot, I felt that this allowed for closer connection with the robot.