Implemented commands into UI
Things still need to be better visualized. The robot will now pick up the closest pickupable.

Also separated the pickup and putdown commands. So now the player first give a pickup, then gives movement commands to destination, before giving the putdown command.

Added direction to connections to highlight flow. Still ugly and needs to be replaced since I just used some random asset.

Added a slider to the nodes to control the amount of time spent on a task. Done because this was more intuitive and easier to control than entering the time manually.

Next task would be to replace the text in the commands with icons. Furthermore the UI still needs to be made prettier, highlighting interactable stuff, etc.